+++ pythainlp/corpus/__init__.py
@@ -44,14 +44,13 @@ __all__: list[str] = [
     "make_safe_directory_name",
 ]
 
-import os
-
 from pythainlp.tools import get_full_data_path, get_pythainlp_path
+from pythainlp.tools.path import safe_path_join
 
 # Remote and local corpus databases
 
 _CORPUS_DIRNAME: str = "corpus"
-_CORPUS_PATH: str = os.path.join(get_pythainlp_path(), _CORPUS_DIRNAME)
+_CORPUS_PATH: str = safe_path_join(get_pythainlp_path(), _CORPUS_DIRNAME)
 _CORPUS_DB_URL: str = "https://pythainlp.org/pythainlp-corpus/db.json"
 
 # filename of local corpus catalog