+++ pythainlp/translate/en_th.py
@@ -10,7 +10,6 @@ Website: https://airesearch.in.th/releases/machine-translation-models/
 
 from __future__ import annotations
 
-import os
 import warnings
 from typing import Optional
 
@@ -29,6 +28,7 @@ except ImportError as e:
     ) from e
 
 from pythainlp.corpus import download, get_corpus_path
+from pythainlp.tools.path import safe_path_join
 
 _EN_TH_MODEL_NAME: str = "scb_1m_en-th_moses"
 # SCB_1M-MT_OPUS+TBASE_en-th_moses-spm_130000-16000_v1.0.tar.gz
@@ -45,7 +45,7 @@ def _get_translate_path(model: str, *path: str) -> str:
     corpus_path = get_corpus_path(model, version="1.0")
     if not corpus_path:
         return ""
-    return os.path.join(corpus_path, *path)
+    return safe_path_join(corpus_path, *path)
 
 
 def _download_install(name: str) -> None: